Trap attempts to sprintf 64-bit ints with %d (part 2)
Rather than printing "NaN" as the outcome of feeding sprintf a too-large
integer with user-supplied format %d (or %x %i %u %o), adjust the format by
inserting __PRI64_PREFIX as defined in <inttypes.h>.

This means that sprintf formats "%d", "%x", etc work for gnuplot "int"
variables regardless of whether the gnuplot executable was built to
use 32-bit or 64-bit integer variables.

+++ src/syscfg.h
@@ -289,7 +289,7 @@ typedef unsigned uintgr_t;	/* no C99 types available */
 #define PLD "%d"
 #define INTGR_MAX INT_MAX
 #define INTGR_MIN INT_MIN
-#define LARGEST_GUARANTEED_NONOVERFLOW INT_MAX
+#define LARGEST_GUARANTEED_NONOVERFLOW (double)(INT_MAX)
 #endif
